Implement Kings & Queens
This puts through a large number of tests for both kings and queens, along with
the base logic for deciding their moves.  At this stage, there remains a number
of outstanding issues:

* En Passant
* Castling
* Check
* Promotion
* Stalemate

Before embarking on these, however, I'd quite like to put together a webassembly
based GUI for playing games.  We are pretty close to that now.